swerve and manual outer arm works

+++ src/main/java/frc/robot/Constants.java
@@ -194,7 +194,7 @@ public static final class OuterArmConstants {
     public static final int outermotor = 13;
     public static final double raiseSpeed = 0.3;
     public static final double lowerSpeed = -0.3;
-    public static final int stopSpeed = 0;
+    public static final double stopSpeed = 0.0;
     public static final double outerP = 10.0;
     public static final double outerI = 0.0;
     public static final double outerD = 0.0;
